# src/backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
    "version": "1.0.0",
    "title": "T-Coffee API"
  },
  "host": "192.168.0.76:3000",
  "basePath": "/api/v1",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/product/{productId}": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Get a product by ID",
        "description": "",
        "operationId": "getProductById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product to return",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "500": {
            "description": "Invalid ID supplied"
          },
          "400": {
            "description": "Please provide product id"
          },
          "404": {
            "description": "Product not found"
          }
        }
      },
      "put": {
        "tags": [
          "product"
        ],
        "summary": "Update an existing product",
        "description": "",
        "operationId": "updateProduct",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product that needs to be updated",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Product object that needs to be updated",
            "schema": {
              "$ref": "#/definitions/updateProduct"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Product updated successfully"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "product"
        ],
        "summary": "Delete a product",
        "description": "",
        "operationId": "deleteProduct",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product that needs to be deleted",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Product deleted successfully"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Product": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "name": {
          "type": "string"
        },
        "price": {
          "type": "number",
          "format": "double"
        },
        "description": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "variants": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Variant"
          }
        },
        "categories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Category"
          }
        }
      }
    },
    "Variant": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "options": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Option"
          }
        }
      }
    },
    "Option": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "additional_price": {
          "type": "number",
          "format": "double"
        },
        "variant_id": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "Category": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "icon": {
          "type": "string"
        }
      }
    },
...
    "updateProduct": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Cappuccino",
          "required": [
            "false"
          ]
        },
        "price": {
          "type": "number",
          "format": "double",
          "example": 39000,
          "required": [
            "false"
          ]
        },
        "description": {
          "type": "string",
          "example": "Cappuccino is a coffee drink that today is typically composed of double espresso and hot milk, with the surface topped with foamed milk.",
          "required": [
            "false"
          ]
        },
        "image": {
          "type": "string",
          "example": "https://www.starbucks.com.vn/media/catalog/product/cache/1/image/9df78eab33525d08d6e5fb8d27136e95/c/a/cappuccino.jpg",
          "required": [
            "false"
          ]
        },
        "variantId": {
          "type": "array",
          "required": [
            "false"
          ],
          "items": {
            "type": "string"
          },
          "example": [
            "size",
            "sugar",
            "topping"
          ]
        },
        "categoryId": {
          "type": "array",
          "required": [
            "false"
          ],
          "items": {
            "type": "string"
          },
          "example": [
            "coffee",
            "drinks"
          ]
        }
      }
    },
...
  }
}